bme review fixes
