+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.Domain/ChargeLinkCommands/IChargeLinkCommandFactory.cs
@@ -14,7 +14,7 @@
 
 using System.Diagnostics.CodeAnalysis;
 using System.Threading.Tasks;
-using GreenEnergyHub.Charges.Domain.CreateLinkCommandEvents;
+using GreenEnergyHub.Charges.Domain.CreateLinkRequest;
 using GreenEnergyHub.Charges.Domain.DefaultChargeLinks;
 
 namespace GreenEnergyHub.Charges.Domain.ChargeLinkCommands
