+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.Infrastructure/Integration/Mappers/CreateDefaultChargeLinksInboundMapper.cs
@@ -14,7 +14,7 @@
 
 using System.Diagnostics.CodeAnalysis;
 using GreenEnergyHub.Charges.Commands;
-using GreenEnergyHub.Charges.Domain.CreateLinkCommandEvents;
+using GreenEnergyHub.Charges.Domain.CreateLinkRequest;
 using GreenEnergyHub.Messaging.Protobuf;
 using GreenEnergyHub.Messaging.Transport;
